bdev/ocssd: zone reset / unmap
Added zone reset and unmap commands.  They're both doing the same thing
(resetting a number of contiguous zones), the only difference is that
zone reset expects the user to pass number of zones to reset, while unmap
requires number of blocks instead (which needs to be aligned to size of
a zone).

Change-Id: Iae4d77c6877b6906231336b13fad1167de149f20
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>