nvme/tcp: Support tx zero copy
If the user asked for it through setting the SPDK_ZC=1 env var and
the socket provider supports that, use the socket tx zero-copy API
of sendmsg(fd , msg, MSG_ZEROCOPY).

Note that this applies for both PDU headers and data.

Since the initiator has transactional communication pattern with the
target, we don't need to care for zero-copy buffer reclaim notifications
from the socket provider, the target response signal that the buffer may
be reused.

fixme: make ZC a controller config (similar to using crc with nvme/tcp)
       and not based on env var. For now one needs to run with SPDK_ZC=1
       to get us use TX zero-copy.

Change-Id: I63af892e74d5700027487afeea505f295bff6afe
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>