bdev/ocssd: initial Open Channel bdev implementation
This patch adds initial Open Channel zoned bdev implementation.  The
bdev will allow to use the zoned API on top of OCSSD devices.

Added the ability to create the OCSSD NVMe controller as well as
creation and deletion of OCSSD bdevs on top of it.  The controller can
be created using the regular NVMe's rpc by specifing 'ocssd' mode.  The
bdevs can be created / deleted via designated rpc calls.  For instance,
the following:

 rpc.py construct_nvme_bdev -b nvme0 -a 0000:00:04.0 -m ocssd -t pcie
 rpc.py bdev_ocssd_create -c nvme0 -b nvme0n1

creates Open Channel controller nvme0 and OCSSD bdev nvme0n1 on top of
it.

Change-Id: I31d271126dba4369ac2eaebd4cc7bdd460e5f808
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>