nvme_rdma: add simple disconnect handling.
This is step one in adding transport agnostic disconnect/reconnect
handling on the NVMe-oF initiator.

Add a flag to the nvme qpair to indicate when the transport qpair fails
or becomes invalid. This will allow the generic layer to respond to this
information as it sees fit.

For now, just add a quick check in the admin completion queue. If that
qpair is failed, try to do a reset and follow the normal path.

This flag is currently only set in response to cm_events in the rdma
transport. If the rdma admin qpair receives a disconnect request, it
will alert the upper layer and the reset will take care of failing up
the stack.

Also add a test to confirm the target disconnect behavior. If run on
master the target_disconnet.sh script will take about 10 minutes because
FIO never gets informed of the controller failure, but as of this patch,
fio properly picks up the disconnect and exits right after the target.

Change-Id: I5bd2c33c04c55c0a6367877ec4d5c27bb234ebd0
Signed-off-by: Seth Howell <seth.howell@intel.com>