bdev/ocssd: update write pointers
Update write pointers during successful write / reset completion.  This
is needed in order to support the append operation.

This patch also introduces the zone busy flag, which makes sure only one
write (or reset) operation can be active at a time.  If a request is
directed to a zone that is already processing another operation, it'll
be deferred by returning -ENOMEM to the bdev layer.

Change-Id: I085438431577e4f0cd2d5d9eff7a220090f21176
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>