Update patch set 2
Patch Set 2:

I feel that instead of fixing what shellcheck is complaining about we probably should leave this one permanently in the "excluded" list and document it.

Bash is sometimes hard to follow when sourcing files so this issues will happen (it's not as clear as in Python for example), and I don't see much benefit of having to write #shellcheck directives everywhere.

Patch-set: 2
Reviewer: Gerrit User 1011256 <1011256@d5d70762-12d0-45a1-890d-524b12d3f735>