Update patch set 2
Patch Set 2: Code-Review+2

There should be an empty line after the first line (the headline) and the rest of the commit message.  Otherwise when you do things like git log --oneline, it will show the entire commit message, not just the first line.  You can also see in the relation chain on the right that it puts the whole message in the link title, not just the first line.

Don't worry about updating the commit message just for that reason - just something to keep in mind for the future.

Patch-set: 2
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2