bdev/nvme: Move bdevs allocation from create_ctrlr() to bdev_nvme_create_bdevs()
This is needed as part of upcoming patches. Today we
create one bdev per namespace. With OCSSD controller
we can create more bdevs than we have namespaces,
so we will need to allocate bdevs when we create them.
We won't know how many of them we will have at the time
of controller creation. Change to allocation of bdevs
itself is done in a separate patch.

Signed-off-by: Maciej Szwed <maciej.szwed@intel.com>
Change-Id: I65ca16c8e3b6b0c35967b4927f52d54043e8ed17