lib/iscsi: Exit flush PDU loop if writev fails by EAGAIN
When the send buffer is full and conn->write_pdu_list is not empty,
if the link goes down at the initiator, writev() continues to return
EAGAIN or EWOULDBLOCK until TCP retransmission timeouts or the link goes
up again. This occupied CPU long time unexpectedly.

So when conn->state is not less than EXITING in iscsi_conn_flush_pdus(),
if writev() returns EWOULDBLOCK or EAGAIN, the flush PDU loop exits immediately.

conn->write_pdu_list is cleared and conn->flush_poller is unregistered
soon after, and so leave them.

  Fixes #971

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I9e1029ffa555e5e43aade53f61e942e98343006d