nvme: move submit_request impl to a private function
This change will pave the way for adding a resubmit flag to the
submit_request function without changing the current API. Ultimately, a
new internal API header will be added which indicates that we want to
resubmit the head of the queued requests. It will encapsulate pulling
off the head of the list, and making sure it is added back to the head
if it cannot be submitted.

This patch series is aimed at preserving the order of qpair entries
when resubmitting queued requests. There are times when we could
possibly run into a condition where the qpair cannot submit the same
number of requests that it reaped. For example, if a controller starts a
reset while a queue is processing completions at the transport level on
a different thread.

Change-Id: Ib20d61d3abaed637c9c305b75081947630190fd4
Signed-off-by: Seth Howell <seth.howell@intel.com>