nvme: make qp_submit_request resubmit aware
This function needs to know whether it is performing a submit or a
resubmit to be able to preserve the order of the queued requests.

The next patch will expose an internal api for attempting to bulk
resubmit requests to the qpair. It will encapsulate pulling from the
front of the queue to resubmit so we don't have to worry about users
(even internal ones) trying to use the resubmit flag incorrectly.

Change-Id: I66605d304834d8255fc82a9a7b71d03bdaa90d27
Signed-off-by: Seth Howell <seth.howell@intel.com>