vhost: move feature fields from backend to vdev structure
This will enable us to make features dynamic, dependent on underlying
backend device.

Dynamic features are used in vhost_blk, but backend structure is
assumed to be static, so we had to call
rte_vhost_driver_enable_features() after we
registered driver with features that backend doesn't support.
This patch makes that configuration simpler by
dynamically setting virtion_features field.

Also, in case of vhost_blk, we don't have to call rte_vhost functions now,
so incapsulation is better and we can now mock all driver
related functions in unit tests if we want to.

Change-Id: Icd76bdd76a3d67ec74e0ac992d8da639beead593
Signed-off-by: Vitaliy Mysak <vitaliy.mysak@intel.com>