nvme_qpair: fix check_enabled.
check_enabled had a few characteristics about it that made it unfriendly
for enabling I/O qpairs after a reset.
1. It didn't check whether we were already in the middle of calling it,
so we could reenter function calls like
nvme_qpair_abort_queued_requests.
2. It was calling nvme_qpair_abort_queued_requests before setting the
enabled flag to true. For applications that submit new I/O in the
completion callback for old I/O, this means you enter an infinite loop
of submitting requests, and then immediately completing them. SO
instead, wait for the qpair to reset, then just submit those requests to
the lower layer.
3. The transport layer resets were being called after the qpair was
enabled. This could cause us to abort requests that were resubmitted
after the qpair was enabled (although this was probably fixed by the
solution to 2)

Change-Id: Ie0b74819a6b16839965bced47c33dec967f725a8
Signed-off-by: Seth Howell <seth.howell@intel.com>