Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> > Patch Set 6: Code-Review-1
> > 
> > I am rethinking this after talking more with Ben.
> > 
> > So I have no doubt that this patch fixes the problem that you've seen.  But I think we need to do some simplification here.  In your GitHub issue, you said this is seen when you disconnect the initiator.  So there was no LOGOUT I assume.  This means that connection must have advanced to EXITING state by detecting that the connection was closed.
> > 
> > In this case, once we have notification that connection was closed, we should not try to write any more data to the socket.  We should just free the buffers.
> > 
> > What do you think?
> 
> I think it's really tough to distinguish between the case where the connection is being closed by the target or is being logged out by the initiator (where you really do want to finish sending data) and the case where the initiator just closed the socket. The best way to tell is to probably try to do a recv() with the MSG_PEEK flag to see if it returns 0 in between each attempt to flush the PDUs.

Hi Jim, Ben,

What I was wondering was that the connection was already in exiting state in this case even if what we did was just link down at the initiator. NOP-In/Out timeout happened before?

So I think what you are saying is how to move the connection to start exiting, and if the connection can start exiting, this patch is fine.

I will talk with my colleague about this aspect.

-Shuhei

Patch-set: 6
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>