lvol: remove clear-method param from lvol create
It's being moved to the delete RPC in a later patch. It
currently doesn't work when specified on create anyway.

The param will still be accepted by the RPC but will error. The
thinking here is that if soneone really wants this option they
should know that they're not getting it via this API.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: Ifa1d722a531e7a5befae4ac7bdb4cc8907256111