lib/iscsi: Move PDU header handling before data buffer allocation
To use zero copy bdev I/O APIs for write I/O, we have to allocate
iSCSI task before allocating data buffer because allocating data
buffer will be changed to spdk_bdev_zcopy_start() call and the
allocated iSCSI task will be passed to the call.

One critical requirement is that we have to read all data for the
current PDU even if handling the header of the PDU fails. For that
purpose, add is_failed flag into PDU and use it.

This change is helpful to unify getting DIF context and allocating
iSCSI task. The next patch will do that.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Ice77b7266af8ac392d9094478523e6e6fe6d131a