sock: Add an API spdk_sock_peek_msg to take a peek at an incoming message
Add an API spdk_sock_peek_msg() to take a peek at an incoming message.
For POSIX socket, it calls recv() with MSG_PEEK and MSG_DONTWAIT.
For VPP socket, it returns always zero.

The first use case of spdk_sock_peek_msg() is to check if the connection
is normal during sending data by tight loop in iSCSI target.

If spdk_sock_peek_msg() returns not positive, exit the loop to send data,
and then cleanup the connection.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I8a826a73a84a4b1ebfb1ac4673bed5324c34773e