lib/iscsi: Check if the connection is normal during flushing PDUs by tight loop
Use the new API spdk_sock_peek_msg() to check if the connection
is normal during flushing PDUs by tight loop. If spdk_sock_peek_msg()
returns non-positive, the connection is very likely to be disconnected,
and so move the connection to EXITING.

Checking if errno is non-zero has similar effect but may be less
accurate and so is not used.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I71ba15bc16529d9b7d871aebfb3df9e76c05f33d