test/vhost: Fix path to integrity_vm.sh
Patch b6af628 while addressing Shellcheck 2088,
broke the pathing to integrity_vm.sh within VM.

Originally tilde was used within quotes so it didn't expand,
since cmd was executed as root the path was correct on VM.

After the change the path was changed to $HOME, making it
actually expand on the host system (not VM). Path was made
incorrect.

Since we already scp to /root/, it shouldn't
be an issue to use that path in explicitly when executing.

This fixes recent nightly run failures.

Change-Id: I2e0ac767cbf77a71841db260e602b568b1f510a1
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>