test/vhost: fix the nightly integrity test.
This test was broken when ~ was changed to $HOME recently. shellcheck
gives a warning for using ~inside of quotes, because it doesn't expand,
but this is actually what we want in this case. When the vm runs the
command, ~ expands in the vm to /root whereas $HOME inside the host
expands to /home/sys_sgci causing the test to fail.

fixes: b6af62867101df732d3a2b601d7dae2baf463763

Change-Id: I29174a00990a427413ea7808b72dc9c9bc13e993
Signed-off-by: Seth Howell <seth.howell@intel.com>