Update patch set 6
Patch Set 6: Code-Review-1

Going back to my comment about separating the bdevs array from the controller.  This is really important.  I really think the common layer shouldn't have to know that OCSSD might create more bdevs than there are namespaces.  We may even have controllers in the future that support key-value namespaces, which won't have bdevs at all.

Plus we need to be prepared for controllers with different kinds of namespaces.

If we need to get this in as-is for now, I can maybe go along with that.  But I want to make sure we are prepared to make the correct long-term changes.  I'd really prefer to get those changes in now so we don't have to redo a bunch of work in this series and Konrad's follow-up series.

Patch-set: 6
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1