[RFC] use posix TCP stack with Seastar
The first patch in this series uses the native DPDK
stack in Seastar.  But using Seastar with the POSIX
stack may still be useful - for example, I used that
to do a lot of the initial development work.

Eventually I will wrap this into some kind of configure
option.

So don't spend time reviewing this patch for now.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I51bf5b387039df8dc066f4938112a012a4edc2ee