Update patch set 1
Patch Set 1: Code-Review-1

I think more details in the commit should be added:

We choose this implementation way, but not queueing the parent nvme request directly in nvme_qpair_submit_request is that:

When the queued parent request is resubmitted, then child nvme request may also be queued if the queue_req list is not empty. Since this child nvme request is firstly time submitted, thus it will be queued in the end of the list. Then there could be no nvme request be submitted to the real transport. Thus there can be no NVMe request completion event later. So it will enter an infinite loop. 

And using this way, all the child nvme request will be submitted in the beginning. Thus those NVMe requests will become "resubmitted I/O" if they are queued. And next time it is executed, it will always be put in the beginning of the queued_req, and it will be finally executed by the transport.

Patch-set: 1
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1